further file decomposition
