+++ frontend/src/types.ts
@@ -1,4 +1,4 @@
-import { availableYears, calculatedSexRecognitionTableColumns, calculatedTableColumnTypes } from "./constants";
+import { availableYears } from "./constants";
 
 export interface DataType {
   key: React.Key;
@@ -76,67 +76,3 @@ export type LineColor = {
 };
 
 export type Sex = "male" | "female";
-
-export type CalculatedSexRecognitionTableColumnTitle = (typeof calculatedSexRecognitionTableColumns)[number]["title"];
-export type CalculatedSexRecognitionTableColumnDataIndex =
-  (typeof calculatedSexRecognitionTableColumns)[number]["dataIndex"];
-
-export type CalculatedSexRecognitionTableRow = {
-  [index in CalculatedSexRecognitionTableColumnDataIndex]: number;
-};
-
-export type CalculatedNoSexRecognitionTableRow = Omit<
-  CalculatedSexRecognitionTableRow,
-  | "menPopulationRussia"
-  | "menMorbidityRussia"
-  | "menIntensiveMorbidityRussia"
-  | "menLowerIntensiveMorbidityRussia"
-  | "menUpperIntensiveMorbidityRussia"
-  | "womenPopulationRussia"
-  | "womenMorbidityRussia"
-  | "womenIntensiveMorbidityRussia"
-  | "womenLowerIntensiveMorbidityRussia"
-  | "womenUpperIntensiveMorbidityRussia"
-  | "menPopulationChosenRegions"
-  | "menMorbidityChosenRegions"
-  | "menIntensiveMorbidityChosenRegions"
-  | "menLowerIntensiveMorbidityChosenRegions"
-  | "menUpperIntensiveMorbidityChosenRegions"
-  | "womenPopulationChosenRegions"
-  | "womenMorbidityChosenRegions"
-  | "womenIntensiveMorbidityChosenRegions"
-  | "womenLowerIntensiveMorbidityChosenRegions"
-  | "womenUpperIntensiveMorbidityChosenRegions"
->;
-
-export type CalculatedTableRow = CalculatedSexRecognitionTableRow | CalculatedNoSexRecognitionTableRow;
-
-export function isCalculatedSexRecognitionTableRow(row: CalculatedTableRow): row is CalculatedSexRecognitionTableRow {
-  return (row as CalculatedSexRecognitionTableRow).menMorbidityRussia !== undefined;
-}
-
-export type PlotlyInputData = {
-  x: string[];
-  y: number[];
-  name: string;
-  error_y: {
-    type: "data" | "constant" | "percent";
-    array: number[];
-    arrayminus: number[];
-    visible: boolean;
-    color: string;
-  };
-  type: "bar";
-};
-
-export type CalculatedTableColumnType = (typeof calculatedTableColumnTypes)[number];
-
-export type CalculationCategoriesType = {
-  Russia: number;
-  menRussia: number;
-  womenRussia: number;
-
-  ChosenRegions: number;
-  menChosenRegions: number;
-  womenChosenRegions: number;
-};
